Move diffing methods from Registry into FcoDiff.py
